Compiletime reduction + build fix (#104)
Fix the latest msvc2022 release compilation bug.

A few compile time reductions:
* when using precompiled headers, use them during edit_tools build as well
* reduce the header dependencies a bit

Replace tabs that crept in, with spaces.

Fix some of gcc/clang/msvc warnings.

Frame callbacks converted from Callable to lambda functions.

+++ core/io/file_access_compressed.cpp
@@ -314,7 +314,7 @@ uint64_t FileAccessCompressed::get_buffer(uint8_t *p_dst, uint64_t p_length) con
         return 0;
     }
 
-    for (int i = 0; i < p_length; i++) {
+    for (uint64_t i = 0; i < p_length; i++) {
 
         p_dst[i] = read_ptr[read_pos];
         read_pos++;
